fix: prevent uint32 overflow in low-RPM fuel/scheduler calculations
fuel_calc.cpp: calc_fuel_pw_us_default_fast — intermediate req_fuel × ve × map
  could reach ~3.8B (> uint32 max) with large engines / small injectors at WOT.
  Cast to uint64_t before the triple multiply.

ecu_sched.cpp: ticks_to_cycle_degrees — denom (tooth_ticks × 120) theoretical
  overflow below 0.27 RPM. Upgrade all arithmetic to uint64_t for safety.

+++ src/engine/fuel_calc.cpp
@@ -381,10 +381,11 @@ uint32_t calc_fuel_pw_us_default_fast(uint8_t ve,
     uint32_t base_pw_us = 0u;
     if (ve != 0u && map_bar_x100 <= 300u) {
         const uint32_t req_fuel_us = default_req_fuel_us();
-        const uint32_t num = req_fuel_us *
-                             static_cast<uint32_t>(ve) *
-                             static_cast<uint32_t>(map_bar_x100);
-        base_pw_us = num / (100u * static_cast<uint32_t>(cfg::g_eng_cfg.map_ref_bar_x100));
+        const uint64_t num = static_cast<uint64_t>(req_fuel_us) *
+                             static_cast<uint64_t>(ve) *
+                             static_cast<uint64_t>(map_bar_x100);
+        base_pw_us = static_cast<uint32_t>(
+            num / (100u * static_cast<uint64_t>(cfg::g_eng_cfg.map_ref_bar_x100)));
         if (base_pw_us > 100000u) {
             base_pw_us = 100000u;
         }
